KREST-5732 Address second round of review comments
